# demo/swagger/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GIN + ENT/SQLX + CASBIN + WIRE",
        "title": "zgo-demo",
        "contact": {
            "name": "suisrc",
            "email": "susirc@outlook.com"
        },
        "license": {},
        "version": "0.0.1"
    },
    "basePath": "/api",
    "paths": {
        "/auth": {
            "get": {
                "description": "授权接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Authorize",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Success"
                        }
                    }
                }
            }
        },
        "/demo/get": {
            "get": {
                "description": "Get",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Demo id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/demo/get1": {
            "get": {
                "description": "Get",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Demo id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/demo/hello": {
            "get": {
                "description": "Hello world",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo"
                ],
                "summary": "Hello",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/demo/set": {
            "post": {
                "description": "Set",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo"
                ],
                "summary": "Set",
                "parameters": [
                    {
                        "description": "Demo Info",
                        "name": "item",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/schema.DemoSet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.ErrorInfo"
                        }
                    }
                }
            }
        },
        "/signin": {
            "post": {
                "description": "登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Signin",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Success"
                        }
                    }
                }
            }
        },
        "/signout": {
            "get": {
                "description": "登出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Signin",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Success"
                        }
                    }
                }
            }
        },
        "/user/hello": {
            "get": {
                "description": "用户接口测试",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "hello",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Success"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helper.ErrorInfo": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "响应数据",
                    "type": "object"
                },
                "errorCode": {
                    "description": "错误代码",
                    "type": "string"
                },
                "errorMessage": {
                    "description": "向用户显示消息",
                    "type": "string"
                },
                "showType": {
                    "description": "错误显示类型：0静音； 1条消息警告； 2消息错误； 4通知； 9页",
                    "type": "integer"
                },
                "success": {
                    "description": "请求成功, false",
                    "type": "boolean"
                },
                "traceId": {
                    "description": "方便进行后端故障排除：唯一的请求ID",
                    "type": "string"
                }
            }
        },
        "helper.Success": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "响应数据",
                    "type": "object"
                },
                "success": {
                    "description": "请求成功, false",
                    "type": "boolean"
                },
                "traceId": {
                    "description": "方便进行后端故障排除：唯一的请求ID",
                    "type": "string"
                }
            }
        },
        "schema.DemoSet": {
            "type": "object",
            "required": [
                "code",
                "name"
            ],
            "properties": {
                "code": {
                    "description": "编号",
                    "type": "string"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}